Additions and functionality for map page
