+++ client/package.json
@@ -11,6 +11,7 @@
     "cors": "^2.8.5",
     "enzyme": "^3.11.0",
     "faker": "^5.5.3",
+    "google-map-react": "^2.1.10",
     "react": "^17.0.2",
     "react-chartjs-2": "^4.0.0",
     "react-dom": "^17.0.2",
@@ -41,5 +42,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@iconify/icons-mdi": "^1.1.42",
+    "@iconify/react": "^3.1.0"
   }
 }
