--- client/src/dashboard-pages/LocationPin.jsx
@@ -1,11 +1,0 @@
-import { Icon } from '@iconify/react'
-import locationIcon from '@iconify/icons-mdi/map-marker'
-
-const LocationPin = ({ text }) => (
-    <div className="pin">
-      <Icon icon={locationIcon} className="pin-icon" />
-      <p className="pin-text">{text}</p>
-    </div>
-  )
-
-  export default Map
