+++ client/src/routing/DashView.jsx
@@ -7,8 +7,10 @@ import Pump from "../dashboard-pages/PumpView";
 import Error from "../dashboard-pages/Error";
 import Contact from "../dashboard-pages/Contact";
 import Messages from "../dashboard-pages/Messages";
+import Map from "../dashboard-pages/Map";
+
 import {
-    Route,
+    Route, 
     Switch,
     //Redirect
   } from "react-router-dom";
@@ -28,6 +30,7 @@ class DashView extends Component {
                             <Route exact path="/pump" component={Pump} />
                             <Route exact path="/contact" component={Contact} />
                             <Route path="/messages" component={Messages} />
+                            <Route exact path="/Map" component={Map} />
                             <Route path="*" component={Error} />
                         </Switch>
                     </div>
