+++ server/.env
@@ -14,3 +14,5 @@ PASS_DB = 1111
 DB_NAME = iwpDB
 
 HOST_URL = http://localhost:3000
+
+MAP = 'AIzaSyCo-vbwLrMwJGi6BHlx5nYbxUA21a5k9kg'
